screen-orientation/lock-basic.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  screen.orientation.unlock();
}, "Test that screen.orientation.unlock() doesn't throw when there is no lock");

test(() => {
  const value = screen.orientation.unlock();
  assert_equals(value, undefined);
}, "Test that screen.orientation.unlock() returns a void value");

promise_test(async t => {
  const value = await screen.orientation.lock('any');
  assert_equals(value, undefined);
}, "Test that screen.orientation.lock returns a promise which will be fulfilled with a void value.");

promise_test(async t => {
  const orientations = [
    'any',
    'portrait',
    'landscape',
    'portrait-secondary',
    'landscape-primary',
    'landscape-secondary',
    'portrait-primary',
  ];
  for (const orientation of orientations) {
    const returnType = screen.orientation.lock(orientation);
    assert_true(returnType instanceof Promise);
    await returnType;
  }
  screen.orientation.unlock();
}, "Test that screen.orientation.lock returns a pending promise.");

promise_test(async t => {
  const preType = screen.orientation.type;
  let newType = "";

  if (preType.indexOf("portrait") != -1) {
    newType = "landscape-primary";
  } else {
    newType = "portrait-primary";
  }

  const p = screen.orientation.lock(newType);
  assert_equals(screen.orientation.type, preType);
  await p;
  assert_equals(screen.orientation.type, newType);
}, "Test that screen.orientation.lock() is actually async");
</script>
