third_party/blink/web_tests/platform/mac-mac10.14/external/wpt/screen-orientation/lock-basic-expected.txt

This is a testharness.js-based test.
PASS Test that screen.orientation.unlock() doesn't throw when there is no lock
PASS Test that screen.orientation.unlock() returns a void value
PASS Test that screen.orientation.lock returns a promise which will be fulfilled with a void value.
FAIL Test that screen.orientation.lock returns a pending promise. assert_true: expected true got false
FAIL Test that screen.orientation.lock() is actually async assert_equals: expected "landscape-primary" but got "portrait-primary"
Harness: the test ran to completion.
